+++ src/database/factories/ContactFactory.php
@@ -42,6 +42,9 @@ class ContactFactory extends Factory
                 'seller_code' => $this->faker->word(),
                 'property_code' => $this->faker->word(),
             ],
+            'idImage' => $this->faker->imageUrl(format: 'jpeg'),
+            'selfieImage' => $this->faker->imageUrl(format: 'jpeg'),
+            'payslipImage' => $this->faker->imageUrl(format: 'jpeg')
         ];
     }
 }