+++ src/tests/Unit/ContactTest.php
@@ -1,8 +1,8 @@
 <?php
 use Illuminate\Foundation\Testing\{RefreshDatabase, WithFaker};
+use Spatie\MediaLibrary\MediaCollections\Models\Media;
 use RLI\Booking\Data\{ContactData, ContactOrderData};
 use RLI\Booking\Models\Contact;
-use Illuminate\Support\Carbon;
 
 uses(RefreshDatabase::class, WithFaker::class);
 
@@ -25,6 +25,9 @@ test('contact has schema attributes', function () {
     expect($contact->employment)->toBeArray();
     expect($contact->co_borrowers)->toBeArray();
     expect($contact->order)->toBeArray();
+    expect($contact->idImage)->toBeInstanceOf(Media::class);
+    expect($contact->selfieImage)->toBeInstanceOf(Media::class);
+    expect($contact->payslipImage)->toBeInstanceOf(Media::class);
 });
 
 test('contact has data', function () {
@@ -43,5 +46,42 @@ test('contact has data', function () {
     expect($data->employment)->toBe($contact->employment);
     expect($data->co_borrowers)->toBe($contact->co_borrowers);
     expect($data->order->toArray())->toBe(ContactOrderData::from($contact->order)->toArray());
+    expect($data->idImage)->toBe($contact->idImage->getUrl());
+    expect($data->selfieImage)->toBe($contact->selfieImage->getUrl());
+    expect($data->payslipImage)->toBe($contact->payslipImage->getUrl());
 });
 
+test('contact can attach media', function () {
+    $idImageUrl = 'https://jn-img.enclaves.ph/Test/idImage.jpg';
+    $selfieImageUrl = 'https://jn-img.enclaves.ph/Test/selfieImage.jpg';
+    $payslipImageUrl = 'https://jn-img.enclaves.ph/Test/payslipImage.jpg';
+    $contact = Contact::factory()->create(['idImage' => null, 'selfieImage' => null, 'payslipImage' => null]);
+    $contact->idImage = $idImageUrl;
+    $contact->selfieImage = $selfieImageUrl;
+    $contact->payslipImage = $payslipImageUrl;
+    $contact->save();
+
+    expect($contact->idImage)->toBeInstanceOf(Media::class);
+    expect($contact->selfieImage)->toBeInstanceOf(Media::class);
+    expect($contact->payslipImage)->toBeInstanceOf(Media::class);
+    expect($contact->idImage->name)->toBe('idImage');
+    expect($contact->idImage->file_name)->toBe('idImage.jpg');
+    expect($contact->payslipImage->file_name)->toBe('payslipImage.jpg');
+    expect($contact->idImage->name)->toBe('idImage');
+    expect($contact->selfieImage->name)->toBe('selfieImage');
+    expect($contact->payslipImage->name)->toBe('payslipImage');
+    expect($contact->idImage->file_name)->toBe('idImage.jpg');
+    expect($contact->selfieImage->file_name)->toBe('selfieImage.jpg');
+    expect($contact->payslipImage->file_name)->toBe('payslipImage.jpg');
+    tap(config('app.url'), function ($host) use ($contact) {
+        expect($contact->idImage->getUrl())->toBe(__(':host/storage/:path', ['host' => $host, 'path' => $contact->idImage->getPathRelativeToRoot()]));
+        expect($contact->selfieImage->getUrl())->toBe(__(':host/storage/:path', ['host' => $host, 'path' => $contact->selfieImage->getPathRelativeToRoot()]));
+        expect($contact->payslipImage->getUrl())->toBe(__(':host/storage/:path', ['host' => $host, 'path' => $contact->payslipImage->getPathRelativeToRoot()]));
+        $contact->idImage->delete();
+    });
+    $contact->selfieImage->delete();
+    $contact->payslipImage->delete();
+    $contact->clearMediaCollection('id-images');
+    $contact->clearMediaCollection('selfie-images');
+    $contact->clearMediaCollection('payslip-images');
+});